moves parse_peers and related functions into tracker.py
