+++ src/bencode.py
@@ -1,15 +1,8 @@
 # encode and decode the Bittorrent serialization format
 # http://www.bittorrent.org/beps/bep_0003.html
 
-from __future__ import annotations
-
 import collections
-from typing import Any, BinaryIO, TYPE_CHECKING
-
-
-if TYPE_CHECKING:
-    import torrent
-from shared_types import PeerAddress, PeerId
+from typing import Any, BinaryIO
 
 
 def parse_string_length(s: BinaryIO, i: bytes = b"") -> int:
@@ -121,41 +114,3 @@ def encode_value(v: int | bytes | str | list[Any] | dict[bytes, Any]) -> bytes:
         return encode_dict(v)
     else:
         raise Exception(f'Unsupported type for bencoding: "{type(v)}"')
-
-
-def parse_compact_peers(raw_bytes: bytes) -> list[PeerAddress]:
-    if (len(raw_bytes) % 6) != 0:
-        raise Exception("Peer list length is not a multiple of 6.")
-    else:
-        peers = []
-        for i in range(0, len(raw_bytes), 6):
-            ip = ".".join(str(x) for x in raw_bytes[i : i + 4]).encode()
-            port = int.from_bytes(raw_bytes[i + 4 : i + 6], byteorder="big")
-            peers.append(PeerAddress(ip=ip, port=port))
-        return peers
-
-
-def replace_ipv6_lookback_with_localhost(address: PeerAddress) -> PeerAddress:
-    if address.ip == b"::1":
-        # return dataclasses.replace(address, ip=b"localhost")
-        return PeerAddress(ip=b"localhost", port=address.port)
-    else:
-        return address
-
-
-def parse_peers(
-    data: bytes | list[collections.OrderedDict[bytes, Any]], torrent_state: torrent.TorrentState
-) -> list[tuple[PeerAddress, PeerId | None]]:
-    peer_list: list[tuple[PeerAddress, PeerId | None]] = []
-    match data:
-        case bytes():
-            peer_list = [(address, None) for address in parse_compact_peers(data)]
-        case list():
-            peer_list = [
-                (PeerAddress(ip=x[b"ip"], port=x[b"port"]), PeerId(x[b"peer id"])) for x in data
-            ]
-    return [
-        (replace_ipv6_lookback_with_localhost(address), peer_id)
-        for address, peer_id in peer_list
-        if address.port != torrent_state.listening_port
-    ]
