+++ src/tracker.py
@@ -1,4 +1,6 @@
+import collections
 import logging
+from typing import Any
 from urllib import parse
 
 import h11
@@ -7,12 +9,50 @@ import trio
 from config import Config
 import torrent
 import http_stream
+from shared_types import PeerAddress, PeerId
 
 logger = logging.getLogger("tracker")
 
 
 def _int2bytes(i: int) -> bytes:
     return b"%d" % i
+
+
+def parse_compact_peers(raw_bytes: bytes) -> list[PeerAddress]:
+    if (len(raw_bytes) % 6) != 0:
+        raise Exception("Peer list length is not a multiple of 6.")
+    else:
+        peers = []
+        for i in range(0, len(raw_bytes), 6):
+            ip = ".".join(str(x) for x in raw_bytes[i : i + 4]).encode()
+            port = int.from_bytes(raw_bytes[i + 4 : i + 6], byteorder="big")
+            peers.append(PeerAddress(ip=ip, port=port))
+        return peers
+
+
+def _replace_ipv6_lookback_with_localhost(address: PeerAddress) -> PeerAddress:
+    if address.ip == b"::1":
+        return PeerAddress(ip=b"localhost", port=address.port)
+    else:
+        return address
+
+
+def parse_peers(
+    data: bytes | list[collections.OrderedDict[bytes, Any]], *, listening_port: int
+) -> list[tuple[PeerAddress, PeerId | None]]:
+    peer_list: list[tuple[PeerAddress, PeerId | None]] = []
+    match data:
+        case bytes():
+            peer_list = [(address, None) for address in parse_compact_peers(data)]
+        case list():
+            peer_list = [
+                (PeerAddress(ip=x[b"ip"], port=x[b"port"]), PeerId(x[b"peer id"])) for x in data
+            ]
+    return [
+        (_replace_ipv6_lookback_with_localhost(address), peer_id)
+        for address, peer_id in peer_list
+        if address.port != listening_port
+    ]
 
 
 def tracker_request(
